+++ src/test/kotlin/framework/ProjectBuilder.kt
@@ -69,6 +69,12 @@ class ProjectBuilder(private val fixture: JavaCodeInsightTestFixture, private va
         configure: Boolean = true,
         allowAst: Boolean = false,
     ) = file(path, code, ".json", configure, allowAst)
+    fun yml(
+        path: String,
+        @Language("yaml") code: String,
+        configure: Boolean = true,
+        allowAst: Boolean = false,
+    ) = file(path, code, ".yml", configure, allowAst)
 
     inline fun dir(path: String, block: ProjectBuilder.() -> Unit) {
         val oldIntermediatePath = intermediatePath