+++ src/test/resources/com/demonwav/mcdev/platform/bukkit/pluginYmlInspection/SimplePlugin.java
@@ -1,0 +1,5 @@
+package test;
+
+import org.bukkit.plugin.java.JavaPlugin;
+
+public class SimplePlugin extends JavaPlugin {}