+++ src/test/resources/com/demonwav/mcdev/platform/bukkit/pluginYmlReferenceContributor/afterRename/RenamedPlugin.java
@@ -1,0 +1,5 @@
+package simple.plugin;
+
+import org.bukkit.plugin.java.JavaPlugin;
+
+public class RenamedPlugin extends JavaPlugin {}