+++ src/test/resources/com/demonwav/mcdev/platform/bungeecord/pluginYmlInspection/SimplePlugin.java
@@ -1,0 +1,5 @@
+package test;
+
+import net.md_5.bungee.api.plugin.Plugin;
+
+public class SimplePlugin extends Plugin {}